+++ packages/dashboard/src/GuardRoute.tsx
@@ -9,8 +9,9 @@ interface GuardRouteProps {
 }
 
 export default function NoPersonaGuardRoute(props: GuardRouteProps) {
-    const { currentPersona } = PersonaContext.useContainer()
     const { redirectTo, ...rest } = props
+    const { currentPersona } = PersonaContext.useContainer()
 
-    return currentPersona ? <Navigate to={redirectTo ?? '/'} replace /> : <Route {...rest} />
+    return null
+    // return currentPersona ? <Navigate to={redirectTo ?? '/'} replace /> : <Route {...rest} />
 }
