+++ packages/dashboard/src/components/PageFrame/FeaturePromotions/FollowUs.png
