+++ packages/dashboard/src/components/PageFrame/FeaturePromotions/SendLuckyDrop.png
