+++ packages/plugin-infra/src/web3/useWeb3State.ts
@@ -3,7 +3,5 @@ import { useActivatedPluginWeb3State } from '../hooks/useActivatedPluginWeb3Stat
 
 export function useWeb3State(expectedPluginID?: string) {
     const pluginID = usePluginIDContext()
-    const web3State = useActivatedPluginWeb3State(expectedPluginID ?? pluginID)
-    if (!web3State) throw new Error('Failed to locate web3 state.')
-    return web3State
+    return useActivatedPluginWeb3State(expectedPluginID ?? pluginID) ?? {}
 }
