+++ packages/plugins/Flow/src/constants.ts
@@ -4,7 +4,7 @@ import { NetworkType, ProviderType } from '@masknet/web3-shared-flow/types'
 export const PLUGIN_META_KEY = 'com.mask.flow'
 export const PLUGIN_ID = 'com.mask.flow'
 export const PLUGIN_NAME = 'Flow Chain'
-export const PLUGIN_ICON = 'F'
+export const PLUGIN_ICON = '⚙️'
 export const PLUGIN_DESCRIPTION = ''
 export const PLUGIN_NETWORKS: Web3Plugin.NetworkDescriptor[] = [
     {
